Breaks circular dependencies
Moves fileSize and stringToJson to lib/common/utils.ts in order to break circular dependencies
