+++ libs/ui-lib/lib/common/components/newFeatureSupportLevels/utils.ts
@@ -1,10 +1,11 @@
-import { Cluster, stringToJSON, SupportLevel } from '../../api';
+import { Cluster, SupportLevel } from '../../api';
 import { ClusterFeatureUsage, FeatureId, FeatureIdToSupportLevel } from '../../types';
 import {
   NewFeatureSupportLevelData,
   NewFeatureSupportLevelMap,
 } from './NewFeatureSupportLevelContext';
 import { TFunction } from 'i18next';
+import { stringToJSON } from '../../utils';
 
 export const getLimitedFeatureSupportLevels = (
   cluster: Cluster,
