+++ libs/ui-lib/lib/ocm/components/clusterConfiguration/networkConfiguration/NetworkConfigurationTableBase.tsx
@@ -1,12 +1,6 @@
 import * as React from 'react';
 import { sortable } from '@patternfly/react-table';
-import {
-  Cluster,
-  Host,
-  HostsTableActions,
-  selectSchedulableMasters,
-  stringToJSON,
-} from '../../../../common';
+import { Cluster, Host, HostsTableActions, selectSchedulableMasters } from '../../../../common';
 import NetworkingStatus from '../../hosts/NetworkingStatus';
 import { useTranslation } from '../../../../common/hooks/use-translation-wrapper';
 import {
@@ -24,6 +18,7 @@ import { HostDetail } from '../../../../common/components/hosts/HostRowDetail';
 import HostsTable from '../../../../common/components/hosts/HostsTable';
 import { ValidationsInfo } from '../../../../common/types/hosts';
 import { useClusterWizardContext } from '../../clusterWizard/ClusterWizardContext';
+import { stringToJSON } from '../../../../common/utils';
 
 export const networkingStatusColumn = (
   onEditHostname?: HostsTableActions['onEditHost'],
