+++ libs/ui-lib/lib/ocm/components/clusterConfiguration/staticIp/data/fromInfraEnv.ts
@@ -1,4 +1,4 @@
-import { HostStaticNetworkConfig, InfraEnv, stringToJSON } from '../../../../../common';
+import { HostStaticNetworkConfig, InfraEnv } from '../../../../../common';
 import { FORM_VIEW_PREFIX, getProtocolType, getYamlComments } from './nmstateYaml';
 import {
   StaticIpInfo,
@@ -11,6 +11,7 @@ import {
 import { isDummyYaml } from './dummyData';
 import { formDataFromInfraEnvField } from './formDataFromInfraEnvField';
 import { getEmptyFormViewHost } from './emptyData';
+import { stringToJSON } from '../../../../../common/utils';
 
 export const getStaticNetworkConfig = (
   infraEnv: InfraEnv,
