+++ libs/ui-lib/lib/ocm/components/hosts/utils.ts
@@ -1,10 +1,9 @@
 import { saveAs } from 'file-saver';
-import {
+import type {
   AlertsContextType,
   Cluster,
   V2ClusterUpdateParams,
   Host,
-  stringToJSON,
   Inventory,
 } from '../../../common';
 
@@ -12,6 +11,7 @@ import { isInOcm, handleApiError, getApiErrorMessage } from '../../api';
 import { updateCluster } from '../../reducers/clusters';
 import { ClustersService } from '../../services';
 import ClustersAPI from '../../services/apis/ClustersAPI';
+import { stringToJSON } from '../../../common/utils';
 
 export const downloadHostInstallationLogs = async (
   addAlert: AlertsContextType['addAlert'],
